Restructure directory to work as Python module and make minimum changes for it to still work
