--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "sdec"]
-	path = sdec
-	url = git@github.com:SunDevilRocketry/sdec.git
