+++ liquid_engine_gui/plumbing/ball_valve.py
@@ -12,7 +12,7 @@
 ###############################################################
 # Standard Imports                                            #
 ###############################################################
-import component_template as SDR_component_template
+from liquid_engine_gui.plumbing import component_template as SDR_component_template
 
 
 class Ball_Valve(SDR_component_template.Component):
