+++ liquid_engine_gui/plumbing/orifice.py
@@ -12,7 +12,7 @@
 ###############################################################
 # Standard Imports                                            #
 ###############################################################
-import component_template as SDR_component_template
+from liquid_engine_gui.plumbing import component_template as SDR_component_template
 
 
 class Orifice(SDR_component_template.Component):
