+++ liquid_engine_gui/plumbing/pipe.py
@@ -18,7 +18,7 @@ from tkinter import *
 ###############################################################
 # Project Imports                                             #
 ###############################################################
-import component_template as SDR_component_template
+from liquid_engine_gui.plumbing import component_template as SDR_component_template
 
 
 ###############################################################
