+++ liquid_engine_gui/plumbing/solenoid.py
@@ -12,7 +12,7 @@
 ###############################################################
 # Project Imports                                             #
 ###############################################################
-import component_template as SDR_component_template
+from liquid_engine_gui.plumbing import component_template as SDR_component_template
 
 
 class Solenoid(SDR_component_template.Component):
