+++ liquid_engine_gui/sensor.py
@@ -12,7 +12,7 @@
 ####################################################################################
 # Imports                                                                          #
 ####################################################################################
-import controller as SDR_controller
+from sdec import controller as SDR_controller
 
 
 ####################################################################################
