+++ liquid_engine_gui/valve.py
@@ -19,7 +19,7 @@ import tkinter as tk
 ####################################################################################
 # Project Imports                                                                  #
 ####################################################################################
-import buttons as SDR_buttons
+from liquid_engine_gui import buttons as SDR_buttons
 
 
 ####################################################################################
